refactor: 컴포넌트 네이밍 수정 및 시멘틱 태그 적용 #171
